backend/requirements.txt

fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
pymongo==4.5.0
motor==3.3.1
python-jose==3.3.0
httpx==0.25.0
